Cleaned up the concurrency test in PrefixLookupConcurrencyTest, added some comments to it, and ensured it would run for every lookup that can pass its check.
Also fixed a bug with the RadixTree where it was returning the internally shared key on prefix results instead of just the portion related to the key for the given node. This required changing the search result typing from byte[] to ReadOnlyMemory<byte> (which is a bonus, as we don't want to expose the original byte array in a mutable way anyways).