Removed the IndirectTrie from the project. This removes its code, its references in the test projects, and the readme file.
The references and benchmark results to the FlatTrie have also been removed from the readme file, but it will remain in the project pending further review.

rmTrie (the wrapper around the external rmTrie nuget package as well as the nuget package reference) has been removed from the alternatives project, and SimpleTrie was moved into the alternatives project to replace it (since it had almost identical runtime characteristitcs) during testing and benchmarking.